+++ CleverTapSDK/CleverTap.m
@@ -1496,18 +1496,11 @@ static NSMutableArray<CTInAppDisplayViewController*> *pendingNotificationControl
     
     CleverTapLogDebug(self.config.logLevel, @"%@: handling push notification: %@", self, notification);
     
-    // check to see whether the push includes a test in-app notification, if so don't process further
-    if ([self didHandleInAppTestFromPushNotificaton:notification]) return;
-    
-    // check to see whether the push includes a test inbox message, if so don't process further
-    if ([self didHandleInboxMessageTestFromPushNotificaton:notification]) return;
-    
-    // check to see whether the push includes a test display unit, if so don't process further
-    if ([self didHandleDisplayUnitTestFromPushNotificaton:notification]) return;
+    // check to see whether the push includes a test in-app notification, test inbox message or test display unit, if so don't process further
+    if ([self _checkAndHandleTestPushPayload:notification]) return;
     
     // notify application with push notification custom extras
     [self _notifyPushNotificationTapped:notification];
-    
     
     dispatch_async(dispatch_get_main_queue(), ^{
         // determine application state
@@ -1538,6 +1531,31 @@ static NSMutableArray<CTInAppDisplayViewController*> *pendingNotificationControl
     });
 #endif
 }
+
+#if !defined(CLEVERTAP_TVOS)
+- (BOOL)_checkAndHandleTestPushPayload:(NSDictionary *)notification {
+    if (notification[@"wzrk_inapp"] || notification[@"wzrk_inbox"] || notification[@"wzrk_adunit"]) {
+        // remove unknown json attributes
+        NSMutableDictionary *testPayload = [NSMutableDictionary new];
+        for (NSString *key in [notification allKeys]) {
+            if ([CTUtils doesString:key startWith:CLTAP_NOTIFICATION_TAG] || [CTUtils doesString:key startWith:CLTAP_NOTIFICATION_TAG_SECONDARY]) {
+                testPayload[key] = notification[key];
+            }
+        }
+        if ([self didHandleInAppTestFromPushNotificaton:testPayload]) {
+            return YES;
+        } else if ([self didHandleInboxMessageTestFromPushNotificaton:testPayload]) {
+            return YES;
+        } else if ([self didHandleDisplayUnitTestFromPushNotificaton:testPayload]) {
+            return YES;
+        } else {
+            CleverTapLogDebug(self.config.logLevel, @"%@: unable to handle test payload in the push notification: %@", self, notification);
+            return NO;
+        }
+    }
+    return NO;
+}
+#endif
 
 - (void)_notifyPushNotificationTapped:(NSDictionary *)notification {
     if (self.pushNotificationDelegate && [self.pushNotificationDelegate respondsToSelector:@selector(pushNotificationTappedWithCustomExtras:)]) {
@@ -2095,7 +2113,7 @@ static NSMutableArray<CTInAppDisplayViewController*> *pendingNotificationControl
         }
     }
     [self saveARP:update];
-    [self processDiscardedEventsRequest:arp];
+    [self processDiscardedEventsRequest:update];
     [self.productConfig updateProductConfigWithOptions:[self _setProductConfig:arp]];
 }