+++ CleverTapSDK/Inbox/controllers/CTInboxController.m
@@ -87,8 +87,8 @@ static NSManagedObjectContext *privateContext;
         CleverTapLogStaticInternal(@"%@: updating messages: %@", self.user, messages);
         BOOL haveUpdates = [self.user updateMessages:messages forContext:privateContext];
         if (haveUpdates) {
-            [self notifyUpdate];
             [self _save];
+            [self notifyUpdate];
         }
     }];
 }
@@ -104,8 +104,8 @@ static NSManagedObjectContext *privateContext;
         CTMessageMO *message = [self _messageForId:messageId];
         if (message) {
             [message setValue:@YES forKey:@"isRead"];
-            [self notifyUpdate];
             [self _save];
+            [self notifyUpdate];
         }
     }];
 }
@@ -191,8 +191,8 @@ static NSManagedObjectContext *privateContext;
         for (CTMessageMO *msg in messages) {
             [privateContext deleteObject:msg];
         }
-        [self notifyUpdate];
         [self _save];
+        [self notifyUpdate];
     }];
 }