+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+### [Version 3.9.2](https://github.com/CleverTap/clevertap-ios-sdk/releases/tag/3.9.2) (February 5, 2021)
+
+##### Fixed
+- Removes unknown JSON attributes while handling Test In-App Notification, Test App Inbox or Test Display Unit
+- Makes `model` property `atomic` (thread-safe)
+- Minor Performance improvements
+
 ### [Version 3.9.1](https://github.com/CleverTap/clevertap-ios-sdk/releases/tag/3.9.1) (October 8, 2020)
 
 ##### Added